Added self test case JPP_0400
