+++ test/component_test.py
@@ -22,8 +22,8 @@
 #
 # --------------------------------------------------------------------------------------------------------------
 #
-VERSION      = "0.47.0"
-VERSION_DATE = "23.10.2024"
+VERSION      = "0.48.0"
+VERSION_DATE = "25.10.2024"
 #
 # --------------------------------------------------------------------------------------------------------------
 #TM***
