+++ test/pytest/pytestfiles/test_21_SELF_ASSIGNMENTS_GOODCASE.py
@@ -14,11 +14,11 @@
 #  limitations under the License.
 # --------------------------------------------------------------------------------------------------------------
 #
-# test_13_CYCLIC_IMPORTS_BADCASE.py
+# test_21_SELF_ASSIGNMENTS_GOODCASE.py
 #
 # XC-CT/ECA3-Queckenstedt
 #
-# 09.01.2024 - 11:50:58
+# 25.10.2024 - 20:21:51
 #
 # --------------------------------------------------------------------------------------------------------------
 
@@ -27,22 +27,14 @@ from pytestlibs.CExecute import CExecute
 
 # --------------------------------------------------------------------------------------------------------------
 
-class Test_CYCLIC_IMPORTS_BADCASE:
+class Test_SELF_ASSIGNMENTS_GOODCASE:
 
 # --------------------------------------------------------------------------------------------------------------
-   # Expected: No values are returned, and JsonPreprocessor throws an exception
+   # Expected: JsonPreprocessor returns expected value
    @pytest.mark.parametrize(
-      "Description", ["JSON file with cyclic imports (JSON file imports itself)",]
+      "Description", ["JSON file with self assignments of strings, lists and dictionaries",]
    )
-   def test_JPP_1150(self, Description):
-      nReturn = CExecute.Execute("JPP_1150")
-      assert nReturn == 0
-# --------------------------------------------------------------------------------------------------------------
-   # Expected: No values are returned, and JsonPreprocessor throws an exception
-   @pytest.mark.parametrize(
-      "Description", ["JSON file with cyclic imports (JSON file imports another file, that is already imported)",]
-   )
-   def test_JPP_1151(self, Description):
-      nReturn = CExecute.Execute("JPP_1151")
+   def test_JPP_1800(self, Description):
+      nReturn = CExecute.Execute("JPP_1800")
       assert nReturn == 0
 # --------------------------------------------------------------------------------------------------------------
