+++ test/pytest/pytestfiles/test_24_PARAMETER_SCOPE_BADCASE.py
@@ -14,11 +14,11 @@
 #  limitations under the License.
 # --------------------------------------------------------------------------------------------------------------
 #
-# test_23_PARAMETER_SCOPE_BADCASE.py
+# test_24_PARAMETER_SCOPE_BADCASE.py
 #
 # XC-CT/ECA3-Queckenstedt
 #
-# 23.10.2024 - 19:41:03
+# 25.10.2024 - 20:21:51
 #
 # --------------------------------------------------------------------------------------------------------------
